+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     -   [is_double](./docs/variables/is_double.md)
     -   [is_float](./docs/variables/is_float.md)
     -   [is_int](./docs/variables/is_int.md)
+    -   [is_integer](./docs/variables/is_integer.md)
     -   [is_object](./docs/variables/is_object.md)
 
 -   Добавлена функция [array](./docs_depricated/funcref/vartype/array/func/array.md) в модуль